refactor: websocket 아톰 export 방식 변경
1. 기존처럼 import 후 export 하지 않고 바로 export로 내보내기를 합니다.
